move directional evaluation into stronge wolfe
